separate code within pad_decoder.[cpp|h] in different files
This becomes done to have better overview about the present PAD decoding types, to have all in one file becomes for one from outside more hard to overview.
