+++ src/dsp_dab/pad_decoder.cpp
@@ -35,9 +35,9 @@ void CPADDecoder::Reset()
 }
 
 void CPADDecoder::Process(const uint8_t* xpad_data,
-                         size_t xpad_len,
-                         bool exact_xpad_len,
-                         const uint8_t* fpad_data)
+                          size_t xpad_len,
+                          bool exact_xpad_len,
+                          const uint8_t* fpad_data)
 {
   // undo reversed byte order + trim long MP2 frames
   size_t used_xpad_len = std::min(xpad_len, sizeof(m_xpad));
@@ -221,276 +221,4 @@ void CPADDecoder::Process(const uint8_t* xpad_data,
   // set last CI
   m_lastXPAD_CI.len = xpad_offset;
   m_lastXPAD_CI.type = xpad_ci_type_continued;
-}
-
-
-// --- CDataGroup -----------------------------------------------------------------
-CDataGroup::CDataGroup(size_t dg_size_max)
-{
-  m_dgRaw.resize(dg_size_max);
-  Reset();
-}
-
-void CDataGroup::Reset()
-{
-  m_dgSize = 0;
-  m_dgSizeNeeded = GetInitialNeededSize();
-}
-
-bool CDataGroup::ProcessDataSubfield(bool start, const uint8_t* data, size_t len)
-{
-  if (start)
-  {
-    Reset();
-  }
-  else
-  {
-    // ignore Data Group continuation without previous start
-    if (m_dgSize == 0)
-      return false;
-  }
-
-  // abort, if needed size already reached (except needed size not yet set)
-  if (m_dgSize >= m_dgSizeNeeded)
-    return false;
-
-  // abort, if maximum size already reached
-  if (m_dgSize == m_dgRaw.size())
-    return false;
-
-  // append Data Subfield
-  size_t copy_len = m_dgRaw.size() - m_dgSize;
-  if (len < copy_len)
-    copy_len = len;
-  memcpy(&m_dgRaw[m_dgSize], data, copy_len);
-  m_dgSize += copy_len;
-
-  // abort, if needed size not yet reached
-  if (m_dgSize < m_dgSizeNeeded)
-    return false;
-
-  return DecodeDataGroup();
-}
-
-bool CDataGroup::EnsureDataGroupSize(size_t desired_dg_size)
-{
-  m_dgSizeNeeded = desired_dg_size;
-  return m_dgSize >= m_dgSizeNeeded;
-}
-
-bool CDataGroup::CheckCRC(size_t len)
-{
-  // ensure needed size reached
-  if (m_dgSize < len + CalcCRC::CRCLen)
-    return false;
-
-  uint16_t crc_stored = m_dgRaw[len] << 8 | m_dgRaw[len + 1];
-  uint16_t crc_calced = CalcCRC::CalcCRC_CRC16_CCITT.Calc(&m_dgRaw[0], len);
-  return crc_stored == crc_calced;
-}
-
-
-// --- CDGLIDecoder -----------------------------------------------------------------
-void CDGLIDecoder::Reset()
-{
-  CDataGroup::Reset();
-
-  m_dgliLength = 0;
-}
-
-bool CDGLIDecoder::DecodeDataGroup()
-{
-  // abort on invalid CRC
-  if (!CheckCRC(2))
-  {
-    CDataGroup::Reset();
-    return false;
-  }
-
-  m_dgliLength = (m_dgRaw[0] & 0x3F) << 8 | m_dgRaw[1];
-
-  CDataGroup::Reset();
-
-  //	fprintf(stderr, "CDGLIDecoder: m_dgliLength: %5zu\n", m_dgliLength);
-
-  return true;
-}
-
-size_t CDGLIDecoder::GetDGLILen()
-{
-  size_t result = m_dgliLength;
-  m_dgliLength = 0;
-  return result;
-}
-
-
-// --- CDynamicLabelDecoder -----------------------------------------------------------------
-void CDynamicLabelDecoder::Reset()
-{
-  CDataGroup::Reset();
-
-  m_dl_sr.Reset();
-  m_label.Reset();
-}
-
-bool CDynamicLabelDecoder::DecodeDataGroup()
-{
-  bool command = m_dgRaw[0] & 0x10;
-
-  size_t field_len = 0;
-  bool cmd_remove_label = false;
-
-  // handle command/segment
-  if (command)
-  {
-    switch (m_dgRaw[0] & 0x0F)
-    {
-      case 0x01: // remove label
-        cmd_remove_label = true;
-        break;
-      default:
-        // ignore command
-        CDataGroup::Reset();
-        return false;
-    }
-  }
-  else
-  {
-    field_len = (m_dgRaw[0] & 0x0F) + 1;
-  }
-
-  size_t real_len = 2 + field_len;
-
-  if (!EnsureDataGroupSize(real_len + CalcCRC::CRCLen))
-    return false;
-
-  // abort on invalid CRC
-  if (!CheckCRC(real_len))
-  {
-    CDataGroup::Reset();
-    return false;
-  }
-
-  // on Remove Label command, display empty label
-  if (cmd_remove_label)
-  {
-    m_label.Reset();
-    return true;
-  }
-
-  // create new segment
-  DL_SEG dl_seg;
-  memcpy(dl_seg.prefix, &m_dgRaw[0], 2);
-  dl_seg.chars.insert(dl_seg.chars.begin(), m_dgRaw.begin() + 2, m_dgRaw.begin() + 2 + field_len);
-
-  CDataGroup::Reset();
-
-  //	fprintf(stderr, "CDynamicLabelDecoder: segnum %d, toggle: %s, chars_len: %2d%s\n", dl_seg.SegNum(), dl_seg.Toggle() ? "Y" : "N", dl_seg.chars.size(), dl_seg.Last() ? " [LAST]" : "");
-
-  // try to add segment
-  if (!m_dl_sr.AddSegment(dl_seg))
-    return false;
-
-  // append new m_label
-  m_label.raw = m_dl_sr.label_raw;
-  m_label.charset = m_dl_sr.dl_segs[0].prefix[1] >> 4;
-  return true;
-}
-
-
-// --- DL_SEG_REASSEMBLER -----------------------------------------------------------------
-void DL_SEG_REASSEMBLER::Reset()
-{
-  dl_segs.clear();
-  label_raw.clear();
-}
-
-bool DL_SEG_REASSEMBLER::AddSegment(DL_SEG& dl_seg)
-{
-  dl_segs_t::const_iterator it;
-
-  // if there are already segments with other toggle value in cache, first clear it
-  it = dl_segs.cbegin();
-  if (it != dl_segs.cend() && it->second.Toggle() != dl_seg.Toggle())
-    dl_segs.clear();
-
-  // if the segment is already there, abort
-  it = dl_segs.find(dl_seg.SegNum());
-  if (it != dl_segs.cend())
-    return false;
-
-  // add segment
-  dl_segs[dl_seg.SegNum()] = dl_seg;
-
-  // check for complete m_label
-  return CheckForCompleteLabel();
-}
-
-bool DL_SEG_REASSEMBLER::CheckForCompleteLabel()
-{
-  dl_segs_t::const_iterator it;
-
-  // check if all segments are in cache
-  int segs = 0;
-  for (int i = 0; i < 8; i++)
-  {
-    it = dl_segs.find(i);
-    if (it == dl_segs.cend())
-      return false;
-
-    segs++;
-
-    if (it->second.Last())
-      break;
-
-    if (i == 7)
-      return false;
-  }
-
-  // append complete m_label
-  label_raw.clear();
-  for (int i = 0; i < segs; i++)
-    label_raw.insert(label_raw.end(), dl_segs[i].chars.begin(), dl_segs[i].chars.end());
-
-  //	std::string m_label((const char*) &label_raw[0], label_raw.size());
-  //	fprintf(stderr, "DL_SEG_REASSEMBLER: new m_label: '%s'\n", m_label.c_str());
-  return true;
-}
-
-
-// --- CMOTDecoder -----------------------------------------------------------------
-void CMOTDecoder::Reset()
-{
-  CDataGroup::Reset();
-
-  m_motLength = 0;
-}
-
-bool CMOTDecoder::DecodeDataGroup()
-{
-  // ignore too short lens
-  if (m_motLength < CalcCRC::CRCLen)
-    return false;
-
-  // only DGs with CRC are supported here!
-
-  // abort on invalid CRC
-  if (!CheckCRC(m_motLength - CalcCRC::CRCLen))
-  {
-    CDataGroup::Reset();
-    return false;
-  }
-
-  CDataGroup::Reset();
-
-  //	fprintf(stderr, "CMOTDecoder: m_motLength: %5zu\n", m_motLength);
-
-  return true;
-}
-
-std::vector<uint8_t> CMOTDecoder::GetMOTDataGroup()
-{
-  std::vector<uint8_t> result(m_motLength);
-  memcpy(&result[0], &m_dgRaw[0], m_motLength);
-  return result;
 }
