+++ src/dsp_dab/pad_decoder.h
@@ -19,6 +19,9 @@
 #pragma once
 
 #include "mot_manager.h"
+#include "pad_decoder_dgli.h"
+#include "pad_decoder_dynamiclabelsegment.h"
+#include "pad_decoder_mot.h"
 #include "tools.h"
 
 #include <atomic>
@@ -28,48 +31,6 @@
 #include <string.h>
 #include <string>
 #include <vector>
-
-
-// --- DL_SEG -----------------------------------------------------------------
-struct DL_SEG
-{
-  uint8_t prefix[2];
-  std::vector<uint8_t> chars;
-
-  bool Toggle() const { return prefix[0] & 0x80; }
-  bool First() const { return prefix[0] & 0x40; }
-  bool Last() const { return prefix[0] & 0x20; }
-  int SegNum() const { return First() ? 0 : ((prefix[1] & 0x70) >> 4); }
-};
-
-typedef std::map<int, DL_SEG> dl_segs_t;
-
-
-// --- DL_SEG_REASSEMBLER -----------------------------------------------------------------
-struct DL_SEG_REASSEMBLER
-{
-  dl_segs_t dl_segs;
-  std::vector<uint8_t> label_raw;
-
-  bool AddSegment(DL_SEG& dl_seg);
-  bool CheckForCompleteLabel();
-  void Reset();
-};
-
-
-// --- DL_STATE -----------------------------------------------------------------
-struct DL_STATE
-{
-  std::vector<uint8_t> raw;
-  int charset;
-
-  DL_STATE() { Reset(); }
-  void Reset()
-  {
-    raw.clear();
-    charset = -1;
-  }
-};
 
 
 // --- XPAD_CI -----------------------------------------------------------------
@@ -96,85 +57,6 @@ struct XPAD_CI
 };
 
 typedef std::list<XPAD_CI> xpad_cis_t;
-
-
-// --- CDataGroup -----------------------------------------------------------------
-class CDataGroup
-{
-public:
-  CDataGroup(size_t dg_size_max);
-  virtual ~CDataGroup() = default;
-
-  bool ProcessDataSubfield(bool start, const uint8_t* data, size_t len);
-
-protected:
-  std::vector<uint8_t> m_dgRaw;
-  size_t m_dgSize;
-  size_t m_dgSizeNeeded;
-
-  virtual size_t GetInitialNeededSize() { return 0; }
-  virtual bool DecodeDataGroup() = 0;
-  bool EnsureDataGroupSize(size_t desired_dg_size);
-  bool CheckCRC(size_t len);
-  void Reset();
-};
-
-
-// --- CDGLIDecoder -----------------------------------------------------------------
-class CDGLIDecoder : public CDataGroup
-{
-public:
-  CDGLIDecoder() : CDataGroup(2 + CalcCRC::CRCLen) { Reset(); }
-
-  void Reset();
-
-  size_t GetDGLILen();
-
-private:
-  size_t m_dgliLength;
-
-  size_t GetInitialNeededSize() override { return 2 + CalcCRC::CRCLen; } // DG len + CRC
-  bool DecodeDataGroup() override;
-};
-
-
-// --- CDynamicLabelDecoder -----------------------------------------------------------------
-class CDynamicLabelDecoder : public CDataGroup
-{
-public:
-  CDynamicLabelDecoder() : CDataGroup(2 + 16 + CalcCRC::CRCLen) { Reset(); }
-
-  void Reset();
-
-  DL_STATE GetLabel() { return m_label; }
-
-private:
-  DL_SEG_REASSEMBLER m_dl_sr;
-  DL_STATE m_label;
-
-  size_t GetInitialNeededSize() override  { return 2 + CalcCRC::CRCLen; } // at least prefix + CRC
-  bool DecodeDataGroup() override;
-};
-
-
-// --- CMOTDecoder -----------------------------------------------------------------
-class CMOTDecoder : public CDataGroup
-{
-public:
-  CMOTDecoder() : CDataGroup(16384) { Reset(); } // = 2^14
-
-  void Reset();
-
-  void SetLen(size_t mot_len) { m_motLength = mot_len; }
-
-  std::vector<uint8_t> GetMOTDataGroup();
-
-private:
-  size_t m_motLength;
-
-  size_t GetInitialNeededSize() override { return m_motLength; } // MOT len + CRC (or zero!)
-  bool DecodeDataGroup() override;
-};
 
 
 // --- IPADDecoderObserver -----------------------------------------------------------------
